[AIRFLOW-5299] Move SQLToGCS to core
For more information check AIP-21.
