+++ tests/operators/test_mssql_to_gcs.py
@@ -19,7 +19,7 @@
 
 import unittest
 
-from airflow.contrib.operators.mssql_to_gcs import \
+from airflow.operators.mssql_to_gcs import \
     MsSqlToGoogleCloudStorageOperator
 from tests.compat import mock
 
@@ -62,8 +62,8 @@ class TestMsSqlToGoogleCloudStorageOperator(unittest.TestCase):
         self.assertEqual(op.bucket, BUCKET)
         self.assertEqual(op.filename, JSON_FILENAME)
 
-    @mock.patch('airflow.contrib.operators.mssql_to_gcs.MsSqlHook')
-    @mock.patch('airflow.contrib.operators.sql_to_gcs.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.mssql_to_gcs.MsSqlHook')
+    @mock.patch('airflow.operators.sql_to_gcs.GoogleCloudStorageHook')
     def test_exec_success_json(self, gcs_hook_mock_class, mssql_hook_mock_class):
         """Test successful run of execute function for JSON"""
         op = MsSqlToGoogleCloudStorageOperator(
@@ -94,8 +94,8 @@ class TestMsSqlToGoogleCloudStorageOperator(unittest.TestCase):
         mssql_hook_mock_class.assert_called_once_with(mssql_conn_id=MSSQL_CONN_ID)
         mssql_hook_mock.get_conn().cursor().execute.assert_called_once_with(SQL)
 
-    @mock.patch('airflow.contrib.operators.mssql_to_gcs.MsSqlHook')
-    @mock.patch('airflow.contrib.operators.sql_to_gcs.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.mssql_to_gcs.MsSqlHook')
+    @mock.patch('airflow.operators.sql_to_gcs.GoogleCloudStorageHook')
     def test_file_splitting(self, gcs_hook_mock_class, mssql_hook_mock_class):
         """Test that ndjson is split by approx_max_file_size_bytes param."""
         mssql_hook_mock = mssql_hook_mock_class.return_value
@@ -125,8 +125,8 @@ class TestMsSqlToGoogleCloudStorageOperator(unittest.TestCase):
             approx_max_file_size_bytes=len(expected_upload[JSON_FILENAME.format(0)]))
         op.execute(None)
 
-    @mock.patch('airflow.contrib.operators.mssql_to_gcs.MsSqlHook')
-    @mock.patch('airflow.contrib.operators.sql_to_gcs.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.mssql_to_gcs.MsSqlHook')
+    @mock.patch('airflow.operators.sql_to_gcs.GoogleCloudStorageHook')
     def test_schema_file(self, gcs_hook_mock_class, mssql_hook_mock_class):
         """Test writing schema files."""
         mssql_hook_mock = mssql_hook_mock_class.return_value
